feat: change files extentions to tsx
