+++ src/components/auth/Copyright.tsx
