+++ src/components/projects/ProjectDetails.tsx
